+++ hive/readme/usage.dart
@@ -1,0 +1,8 @@
+import 'package:hive_ce/hive.dart';
+
+void example() {
+  final box = Hive.box('myBox');
+  box.put('name', 'David');
+  final name = box.get('name');
+  print('Name: $name');
+}