+++ hive_generator/lib/src/type_adapter_generator.dart
@@ -4,7 +4,8 @@ import 'package:hive_ce/hive.dart';
 import 'package:hive_ce_generator/src/builder.dart';
 import 'package:hive_ce_generator/src/class_builder.dart';
 import 'package:hive_ce_generator/src/enum_builder.dart';
-import 'package:hive_ce_generator/src/helper.dart';
+import 'package:hive_ce_generator/src/helper/helper.dart';
+import 'package:hive_ce_generator/src/model/hive_schema.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:source_helper/source_helper.dart';
 
@@ -16,24 +17,41 @@ class TypeAdapterGenerator extends GeneratorForAnnotation<HiveType> {
     ConstantReader annotation,
     BuildStep buildStep,
   ) async {
+    final result = generateTypeAdapter(
+      element: element,
+      library: await buildStep.inputLibrary,
+      typeId: readTypeId(annotation),
+      adapterName: readAdapterName(annotation),
+    );
+    return result.content;
+  }
+
+  /// Generate a type adapter with the given information
+  ///
+  /// If this is an incremental update, pass the existing [schema]
+  static GenerateTypeAdapterResult generateTypeAdapter({
+    required Element element,
+    required LibraryElement library,
+    required int typeId,
+    String? adapterName,
+    HiveSchemaType? schema,
+  }) {
     final cls = getClass(element);
-    final library = await buildStep.inputLibrary;
-    final gettersAndSetters = getAccessors(cls, library);
+    final getAccessorsResult =
+        _getAccessors(cls: cls, library: library, schema: schema);
 
-    final getters = gettersAndSetters[0];
-    verifyFieldIndices(getters);
+    final getters = getAccessorsResult.getters;
+    _verifyFieldIndices(getters);
 
-    final setters = gettersAndSetters[1];
-    verifyFieldIndices(setters);
+    final setters = getAccessorsResult.setters;
+    _verifyFieldIndices(setters);
 
-    final typeId = getTypeId(annotation);
-
-    final adapterName = getAdapterName(cls.name, annotation);
+    adapterName ??= generateAdapterName(cls.name);
     final builder = cls.thisType.isEnum
         ? EnumBuilder(cls, getters)
         : ClassBuilder(cls, getters, setters);
 
-    return '''
+    final content = '''
     class $adapterName extends TypeAdapter<${cls.name}> {
       @override
       final int typeId = $typeId;
@@ -59,20 +77,41 @@ class TypeAdapterGenerator extends GeneratorForAnnotation<HiveType> {
               typeId == other.typeId;
     }
     ''';
+
+    return GenerateTypeAdapterResult(content, getAccessorsResult.schema);
   }
 
   /// TODO: Document this!
-  Set<String> getAllAccessorNames(InterfaceElement cls) {
-    final accessorNames = <String>{};
+  static Set<String> _getAllAccessorNames(InterfaceElement cls) {
+    final isEnum = cls.thisType.isEnum;
+    final constructorFields =
+        getConstructor(cls).parameters.map((it) => it.name).toSet();
 
+    final accessorNames = <String>{};
     final supertypes = cls.allSupertypes.map((it) => it.element);
     for (final type in [cls, ...supertypes]) {
+      // Ignore Object base members
+      if (const TypeChecker.fromRuntime(Object).isExactly(type)) continue;
+
       for (final accessor in type.accessors) {
+        // Ignore any non-enum accessors on enums
+        if (isEnum && !accessor.returnType.isEnum) continue;
+
+        // Ignore static fields on non-enums (enum values are considered static)
+        if (!isEnum && accessor.isStatic) continue;
+
+        // Ignore getters without setters on classes
+        if (!isEnum &&
+            accessor.isGetter &&
+            accessor.correspondingSetter == null &&
+            !constructorFields.contains(accessor.name)) continue;
+
+        final name = accessor.name;
         if (accessor.isSetter) {
-          final name = accessor.name;
+          // Remove '=' from setter name
           accessorNames.add(name.substring(0, name.length - 1));
         } else {
-          accessorNames.add(accessor.name);
+          accessorNames.add(name);
         }
       }
     }
@@ -81,32 +120,49 @@ class TypeAdapterGenerator extends GeneratorForAnnotation<HiveType> {
   }
 
   /// TODO: Document this!
-  List<List<AdapterField>> getAccessors(
-    InterfaceElement cls,
-    LibraryElement library,
-  ) {
-    final accessorNames = getAllAccessorNames(cls);
+  static _GetAccessorsResult _getAccessors({
+    required InterfaceElement cls,
+    required LibraryElement library,
+    HiveSchemaType? schema,
+  }) {
+    final accessorNames = _getAllAccessorNames(cls);
 
-    final constructor = cls.constructors.firstWhere((e) => e.name.isEmpty);
+    final constr = getConstructor(cls);
     final parameterDefaults = {
-      for (final param in constructor.parameters)
-        param.name: param.defaultValueCode,
+      for (final param in constr.parameters) param.name: param.defaultValueCode,
     };
 
+    var nextIndex = schema?.nextIndex ?? 0;
+    final newSchemaFields = <String, HiveSchemaField>{};
     AdapterField? accessorToField(PropertyAccessorElement? element) {
       if (element == null) return null;
 
       final annotation =
           getHiveFieldAnn(element.variable2) ?? getHiveFieldAnn(element);
-      if (annotation == null) return null;
+      if (schema == null && annotation == null) return null;
 
       final field = element.variable2!;
+      final name = field.name;
+      final int index;
+      if (schema != null) {
+        // Only generate one id per field name
+        index = schema.fields[name]?.index ??
+            newSchemaFields[name]?.index ??
+            nextIndex++;
+      } else if (annotation != null) {
+        index = annotation.index;
+      } else {
+        // This should be impossible
+        throw HiveError('No index found');
+      }
+
+      newSchemaFields[name] = HiveSchemaField(index: index);
       return AdapterField(
-        annotation.index,
-        field.name,
+        index,
+        name,
         field.type,
-        annotation.defaultValue,
-        parameterDefaults[field.name],
+        annotation?.defaultValue,
+        parameterDefaults[name],
       );
     }
 
@@ -123,16 +179,25 @@ class TypeAdapterGenerator extends GeneratorForAnnotation<HiveType> {
       if (setterField != null) setters.add(setterField);
     }
 
-    return [getters, setters];
+    // Sort by index for deterministic output
+    getters.sort((a, b) => a.index.compareTo(b.index));
+    setters.sort((a, b) => a.index.compareTo(b.index));
+    final newSchema = schema?.copyWith(
+      nextIndex: nextIndex,
+      fields: Map.fromEntries(
+        newSchemaFields.entries.toList()
+          ..sort((a, b) => a.value.index.compareTo(b.value.index)),
+      ),
+    );
+    return _GetAccessorsResult(getters, setters, newSchema);
   }
 
   /// TODO: Document this!
-  void verifyFieldIndices(List<AdapterField> fields) {
+  static void _verifyFieldIndices(List<AdapterField> fields) {
     for (final field in fields) {
-      check(
-        field.index >= 0 && field.index <= 255,
-        'Field numbers can only be in the range 0-255.',
-      );
+      if (field.index < 0 || field.index > 255) {
+        throw 'Field numbers can only be in the range 0-255.';
+      }
 
       for (final otherField in fields) {
         if (otherField == field) continue;
@@ -145,13 +210,24 @@ class TypeAdapterGenerator extends GeneratorForAnnotation<HiveType> {
       }
     }
   }
+}
 
-  /// TODO: Document this!
-  int getTypeId(ConstantReader annotation) {
-    check(
-      !annotation.read('typeId').isNull,
-      'You have to provide a non-null typeId.',
-    );
-    return annotation.read('typeId').intValue;
-  }
+class _GetAccessorsResult {
+  final List<AdapterField> getters;
+  final List<AdapterField> setters;
+  final HiveSchemaType? schema;
+
+  const _GetAccessorsResult(this.getters, this.setters, this.schema);
+}
+
+/// Result of [TypeAdapterGenerator.generateTypeAdapter]
+class GenerateTypeAdapterResult {
+  /// The generated content
+  final String content;
+
+  /// The generated schema
+  final HiveSchemaType? schema;
+
+  /// Constructor
+  const GenerateTypeAdapterResult(this.content, this.schema);
 }